Move NavCard to own component
